+++ lib/ocrypto/ec_key_pair_test.go
@@ -94,17 +94,33 @@ func TestNanoTDFRewrapKeyGenerate(t *testing.T) {
 	digest := sha256.New()
 	digest.Write([]byte("L1L"))
 
-	kasSymmetricKey, err := CalculateHKDF(digest.Sum(nil), kasECDHKey, 32)
+	kasSymmetricKey, err := CalculateHKDF(digest.Sum(nil), kasECDHKey)
 	require.NoError(t, err, "fail to calculate HKDF key")
 
 	sdkECDHKey, err := ComputeECDHKey([]byte(sdkPrivateKeyAsPem), []byte(kasPubKeyAsPem))
 	require.NoError(t, err, "fail to calculate ecdh key")
 
-	sdkSymmetricKey, err := CalculateHKDF(digest.Sum(nil), sdkECDHKey, 32)
+	sdkSymmetricKey, err := CalculateHKDF(digest.Sum(nil), sdkECDHKey)
 	require.NoError(t, err, "fail to calculate HKDF key")
 
 	if string(kasSymmetricKey) != string(sdkSymmetricKey) {
 		t.Fatalf("symmetric keys on both kas and sdk should be same kas:%s sdk:%s",
 			string(kasSymmetricKey), string(sdkSymmetricKey))
+	}
+}
+
+func TestECDSASignature(t *testing.T) {
+	digest := CalculateSHA256([]byte("Virtru"))
+	for _, cvurve := range []ECCMode{ECCModeSecp256r1, ECCModeSecp384r1, ECCModeSecp521r1} {
+		ecKeyPair, err := NewECKeyPair(cvurve)
+		require.NoError(t, err, "fail on NewECKeyPair")
+
+		rBytes, sBytes, err := ComputeECDSASig(digest, ecKeyPair.PrivateKey)
+		require.NoError(t, err, "fail on ComputeECDSASig")
+
+		verify := VerifyECDSASig(digest, rBytes, sBytes, &ecKeyPair.PrivateKey.PublicKey)
+		if verify == false {
+			t.Fatalf("Fail to verify ECDSA Signature")
+		}
 	}
 }