+++ service/cmd/provisionFixtures.go
@@ -26,7 +26,7 @@ var (
 2. Change the 'opentdf.environment.POSTGRES_DB' from 'opentdf' to another value of your choosing (e.g. 'opentdf_local')
 3. Open your root-level 'opentdf.yaml' config
 4. Change the config so the 'db.database' key matches the value of the docker-compose 'POSTGRES_DB' (step 2)
-5. Run 'docker-compose up' to start the connection to the new database
+5. Run 'docker compose up' to start the connection to the new database
 6. Run this command to provision the fixtures data
 
 This command will run the local provision of fixtures data. This will migrate the database to the
@@ -34,7 +34,7 @@ latest version and provision the fixtures. To avoid any pollution of your local 
 to run this command in a clean database. This command is intended for local development and testing purposes only.
 
 ** Teardown or Issues **
-You can clear/recycle your database with 'docker-compose down' and 'docker-compose up' to start fresh.`,
+You can clear/recycle your database with 'docker compose down' and 'docker compose up' to start fresh.`,
 		Run: func(cmd *cobra.Command, _ []string) {
 			configFile, _ := cmd.Flags().GetString(configFileFlag)
 			configKey, _ := cmd.Flags().GetString(configKeyFlag)