+++ service/internal/entitlements/pdp.go
@@ -1,30 +1,43 @@
 package entitlements
 
 import (
+	"encoding/json"
+
 	"github.com/opentdf/platform/protocol/go/authorization"
 	"github.com/opentdf/platform/protocol/go/policy/attributes"
+	"google.golang.org/protobuf/encoding/protojson"
 )
 
 func OpaInput(entity *authorization.Entity, sms map[string]*attributes.GetAttributeValuesByFqnsResponse_AttributeAndValue, ersURL string, authToken string) (map[string]interface{}, error) {
 	// OPA wants this as a generic map[string]interface{} and will not handle
 	// deserializing to concrete structs
 	inputUnstructured := make(map[string]interface{})
+
 	// SubjectMapping
-	inputUnstructured["attribute_mappings"] = sms
-	// Entity
-	ea := make(map[string]interface{})
-	ea["id"] = entity.GetId()
-	switch entity.GetEntityType().(type) {
-	case *authorization.Entity_ClientId:
-		ea["client_id"] = entity.GetClientId()
-	case *authorization.Entity_EmailAddress:
-		ea["email_address"] = entity.GetEmailAddress()
-	case *authorization.Entity_UserName:
-		ea["user_name"] = entity.GetUserName()
-	case *authorization.Entity_Uuid:
-		ea["uuid"] = entity.GetUuid()
+	// convert sms to json string
+	smsJSON := make(map[string]string)
+	for k, v := range sms {
+		attrDefBytes, err := protojson.Marshal(v)
+		if err != nil {
+			return nil, err
+		}
+		smsJSON[k] = string(attrDefBytes)
 	}
-	inputUnstructured["entity"] = ea
+	inputUnstructured["attribute_mappings"] = smsJSON
+
+	// Entity
+	// convert entity to map[string]string
+	eaJSON, err := protojson.Marshal(entity)
+	if err != nil {
+		return nil, err
+	}
+	var eaMap map[string]string
+	err = json.Unmarshal(eaJSON, &eaMap)
+	if err != nil {
+		return nil, err
+	}
+
+	inputUnstructured["entity"] = eaMap
 	inputUnstructured["ers_url"] = ersURL
 	inputUnstructured["auth_token"] = authToken