+++ service/internal/security/crypto_provider.go
@@ -1,6 +1,9 @@
 package security
 
-import "crypto"
+import (
+	"crypto"
+	"crypto/elliptic"
+)
 
 type CryptoProvider interface {
 	RSAPublicKey(keyID string) (string, error)
@@ -8,7 +11,8 @@ type CryptoProvider interface {
 	RSADecrypt(hash crypto.Hash, keyID string, keyLabel string, ciphertext []byte) ([]byte, error)
 
 	ECPublicKey(keyID string) (string, error)
-	GenerateNanoTDFSymmetricKey(ephemeralPublicKeyBytes []byte) ([]byte, error)
+	ECCertificate(keyID string) (string, error)
+	GenerateNanoTDFSymmetricKey(ephemeralPublicKeyBytes []byte, curve elliptic.Curve) ([]byte, error)
 	GenerateEphemeralKasKeys() (any, []byte, error)
 	GenerateNanoTDFSessionKey(privateKeyHandle any, ephemeralPublicKey []byte) ([]byte, error)
 	Close()