+++ lib/ocrypto/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.1.5](https://github.com/opentdf/platform/compare/lib/ocrypto/v0.1.4...lib/ocrypto/v0.1.5) (2024-05-31)
+
+
+### Features
+
+* **sdk:** Support for ECDSA policy binding on both KAS and SDK ([#877](https://github.com/opentdf/platform/issues/877)) ([7baf039](https://github.com/opentdf/platform/commit/7baf03928eb3d29f615359860f9217a69b51c1fe))
+
+## [0.1.4](https://github.com/opentdf/platform/compare/lib/ocrypto/v0.1.3...lib/ocrypto/v0.1.4) (2024-05-29)
+
+
+### Features
+
+* **core:** Allow app specified session keys ([#882](https://github.com/opentdf/platform/issues/882)) ([529fb0e](https://github.com/opentdf/platform/commit/529fb0ec775eca93f8cdd83388eba950a5e81bba))
+* **sdk:** PLAT-3082 nanotdf encrypt ([#744](https://github.com/opentdf/platform/issues/744)) ([6c82536](https://github.com/opentdf/platform/commit/6c8253689ec65e68c2114750c10c501423cbe03c))
+
 ## [0.1.3](https://github.com/opentdf/platform/compare/lib/ocrypto/v0.1.2...lib/ocrypto/v0.1.3) (2024-05-20)